makes upload function to component
