+++ src/components/Cards/CardProfile.js
@@ -1,48 +1,9 @@
 import React from "react";
 
-import {useState} from "react";
-import { useRef } from 'react';
-import axios from 'axios';
-import useFileUpload from 'react-use-file-upload';
-
-
-
-
 // components
+import CardUpload from "components/Cards/CardUpload.js"
 
 export default function CardProfile() {
-
-  
-  const {
-    files,
-    fileNames,
-    fileTypes,
-    totalSize,
-    totalSizeInBytes,
-    handleDragDropEvent,
-    clearAllFiles,
-    createFormData,
-    setFiles,
-    removeFile,
-  } = useFileUpload();
-  
-  const inputRef = useRef();
-  const handleSend = async (e) => {
-    e.preventDefault();
-  
-    const formData = createFormData();
-  
-    try {
-      // axios.post("../data/", formData, {
-      //   'content-type': 'multipart/form-data',
-      // });
-      localStorage.setItem("resume", formData);
-      
-      console.log(localStorage.getItem("resume"));
-    } catch (error) {
-      console.error('Failed to submit files.');
-    }
-  };
 
   return (
     <>
@@ -89,67 +50,7 @@ export default function CardProfile() {
               <div className="text-center">
               Upload Resume
               
-              <div className="form-container justify-center pt-2">
-                      {/* Display the files to be uploaded */}
-                      <div>
-                        <ul>
-                          {fileNames.map((name) => (
-                            <li key={name}>
-                              <span>{name}</span>
-
-                              <span onClick={() => removeFile(name)}>
-                                <i className="fa fa-times" />
-                              </span>
-                            </li>
-                          ))}
-                        </ul>
-
-                        {files.length > 0 && (
-                          <ul>
-                            <li>File types found: {fileTypes.join(', ')}</li>
-                            <li>Total Size: {totalSize}</li>
-                            <li>Total Bytes: {totalSizeInBytes}</li>
-
-                            <li className="clear-all">
-                              <button onClick={() => clearAllFiles()}>Clear All</button>
-                            </li>
-                          </ul>
-                        )}
-                      </div>
-
-                      {/* Provide a drop zone and an alternative button inside it to upload files. */}
-                      <div
-                        // css={Dropzone}
-                        onDragEnter={handleDragDropEvent}
-                        onDragOver={handleDragDropEvent}
-                        onDrop={(e) => {
-                          handleDragDropEvent(e);
-                          setFiles(e, 'a');
-                        }}
-                      >
-                        {/* <p>Drag and drop files here</p> */}
-
-                        <button class="bg-lightBlue-500 active:bg-lightBlue 600 text-white hover:shadow-md shadow tet-xs px-2 py-1 rounded outline-none 
-                        focus:outline-none sm:mr-2 mb-1 ease-linear transition-all duration-150" onClick={() => inputRef.current.click()}>Select files to upload</button>
-
-                        {/* Hide the crappy looking default HTML input */}
-                        <input
-                          ref={inputRef}
-                          type="file"
-                          multiple
-                          style={{ display: 'none' }}
-                          onChange={(e) => {
-                            setFiles(e, 'a');
-                            inputRef.current.value = null;
-                          }}
-                        />
-                        <div className="relative flex flex-col h-16 break-words bg-blueGray-50 w-full mt-2 mb-4 text-blueGray-500"> 
-                        <div className="mt-5">
-                        Drop File Here 
-                          </div>
-                          </div>
-                      </div>
-                      </div>
+              <CardUpload />
                       </div>
                     </div>
             </div>
