+++ src/components/Cards/CardSettings.js
@@ -2,8 +2,6 @@
 import React from "react";
 import {useState} from "react";
 import addPreferences from "../../Functions/addPreferences";
-import { useRef } from 'react';
-import useFileUpload from 'react-use-file-upload';
 
 const CardSettings = () => {
 
@@ -102,36 +100,7 @@ const changeCountry =(e) =>{
 const changeInstitute =(e) =>{
   setInstitute(e.target.value);
 }
-const {
-  files,
-  fileNames,
-  fileTypes,
-  totalSize,
-  totalSizeInBytes,
-  handleDragDropEvent,
-  clearAllFiles,
-  createFormData,
-  setFiles,
-  removeFile,
-} = useFileUpload();
 
-const inputRef = useRef();
-const handleSend = async (e) => {
-  e.preventDefault();
-
-  const formData = createFormData();
-
-  try {
-    // axios.post("../data/", formData, {
-    //   'content-type': 'multipart/form-data',
-    // });
-    localStorage.setItem("resume", formData);
-    
-    console.log(localStorage.getItem("resume"));
-  } catch (error) {
-    console.error('Failed to submit files.');
-  }
-};
   return (
     <>
       <div className="relative flex flex-col min-w-0 break-words w-full mb-6 shadow-lg rounded-lg bg-blueGray-100 border-0">
