+++ src/views/user/Profile.js
@@ -1,39 +1,8 @@
 import React, { useRef } from "react";
-import useFileUpload from 'react-use-file-upload';
 
+import CardUpload from "components/Cards/CardUpload.js"
 
 export default function Profile() {
-  const {
-    files,
-    fileNames,
-    fileTypes,
-    totalSize,
-    totalSizeInBytes,
-    handleDragDropEvent,
-    clearAllFiles,
-    createFormData,
-    setFiles,
-    removeFile,
-  } = useFileUpload();
-  
-  const inputRef = useRef();
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-  
-    const formData = createFormData();
-  
-    try {
-      // axios.post("../data/", formData, {
-      //   'content-type': 'multipart/form-data',
-      // });
-      localStorage.setItem("resume", formData);
-      
-      console.log(localStorage.getItem("resume"));
-    } catch (error) {
-      console.error('Failed to submit files.');
-    }
-  };
-
   return (
     <>
       <main className="profile-page">
@@ -100,69 +69,7 @@ export default function Profile() {
                         className="bg-blueGrey-100 mt-10 uppercase text-black font-bold hover:shadow-md shadow text-xs px-4 py-2 rounded outline-none focus:outline-none sm:mr-2 mb-1 ease-linear transition-all duration-150"
                       >
                         Upload Resume
-                        <div className="form-container">
-                          {/* Display the files to be uploaded */}
-                          <div>
-                            <ul>
-                              {fileNames.map((name) => (
-                              <li key={name}>
-                              <span>{name}</span>
-
-                              <span onClick={() => removeFile(name)}>
-                                <i className="fa fa-times" />
-                              </span>
-                              </li>
-                              ))}
-                            </ul>
-
-                          {files.length > 0 && (
-                            <ul>
-                              <li>File types found: {fileTypes.join(', ')}</li>
-                              <li>Total Size: {totalSize}</li>
-                              <li>Total Bytes: {totalSizeInBytes}</li>
-
-                              <li className="clear-all">
-                                <button onClick={() => clearAllFiles()}>Clear All</button>
-                              </li>
-                            </ul>
-                            )}
-                        </div>
-
-                      {/* Provide a drop zone and an alternative button inside it to upload files. */}
-                      <div
-                        // css={Dropzone}
-                        onDragEnter={handleDragDropEvent}
-                        onDragOver={handleDragDropEvent}
-                        onDrop={(e) => {
-                          handleDragDropEvent(e);
-                          setFiles(e, 'a');
-                        }}
-                      >
-                        {/* <p>Drag and drop files here</p> */}
-
-                        <button class="bg-lightBlue-500 active:bg-lightBlue 600 text-white hover:shadow-md shadow tet-xs px-2 py-1 rounded outline-none 
-                        focus:outline-none sm:mr-2 mb-1 ease-linear transition-all duration-150" onClick={() => inputRef.current.click()}>Select files to upload</button>
-
-                        {/* Hide the crappy looking default HTML input */}
-                        <input
-                          ref={inputRef}
-                          type="file"
-                          multiple
-                          style={{ display: 'none' }}
-                          onChange={(e) => {
-                            setFiles(e, 'a');
-                            inputRef.current.value = null;
-                          }}
-                        />
-                         <div className="relative flex flex-col min-w-0 break-words w-6\/12 mb-0 shadow-lg rounded bg-blueGray-200 border-0">
-                        <div className="relative flex flex-col h-8 break-words bg-blueGray-200 w-6\/12 mt-5 mb-4 text-blueGray-500"> Drop File Here </div>
-                        </div>
-                      </div>
-                    </div>
-                    <div className="submit">
-                      <button class="bg-lightBlue-500 active:bg-lightBlue 600 uppercase text-white font-bold hover:shadow-md shadow tet-xs px-3 py-1
-                      rounded outline-none focus:outline-none sm:mr-2 mb-1 ease-linear transition-all duration-150" onClick={handleSubmit}>Submit</button>
-                    </div>
+                        <CardUpload />
                       </div>
                 </div>
                 <div className="mt-10 py-10 border-t border-blueGray-200 text-center">
